Move definition of ALL
