+++ core/src/View/Kokkos_ViewMapping.hpp
@@ -29,6 +29,7 @@
 #include <impl/Kokkos_Error.hpp>
 #include <impl/Kokkos_Traits.hpp>
 #include <View/Kokkos_ViewTracker.hpp>
+#include <View/Kokkos_ViewTraits.hpp>
 #include <View/Kokkos_ViewCtor.hpp>
 #include <View/Kokkos_ViewAtomic.hpp>
 #include <impl/Kokkos_Tools.hpp>
@@ -41,14 +42,6 @@
 //----------------------------------------------------------------------------
 
 namespace Kokkos {
-
-struct ALL_t {
-  KOKKOS_INLINE_FUNCTION
-  constexpr const ALL_t& operator()() const { return *this; }
-
-  KOKKOS_INLINE_FUNCTION
-  constexpr bool operator==(const ALL_t&) const { return true; }
-};
 
 #ifdef KOKKOS_ENABLE_DEPRECATED_CODE_4
 namespace Impl {
